split Display compontent for smaller pieces
