+++ lib/types/types.ts
@@ -1,7 +1,6 @@
-export default interface Product {
+export interface Product {
   id: number;
   title: string;
-  image: string[];
   images: string[];
   price: number;
   description: string;
